Revert "[EN-20691] Update our SDK fork with the latest upstream changes (#9)"
This reverts commit ab1f31dad8fe2223f9ab852fba07b97212d6801b.

+++ src/main/java/com/eppo/sdk/dto/Allocation.java
@@ -6,6 +6,6 @@ import lombok.Data;
 
 @Data
 public class Allocation {
-  private double percentExposure;
+  private float percentExposure;
   private List<Variation> variations;
 }
